Java Juice Cafe Website
This is a website that I created for the fictional "Java Juice Cafe." This particular website demonstrates the Parallax effect, commonly found in websites today, and allows text and images to scroll on top of a static background. I found this effect to be interesting and mesmerizing to watch and play with whilst coding this website.