Product Customization C++ Script
This is a C++ that asks the user what action they would like to preform from a given list ('S', for shipping fee, 'D', for discount rate, and 'M', for multiplying the quantity of items they would like to purchase by the price of said item. After this action is completed however much the user would like to do it, the user then gets a fully customized receipt based on what they purchased, and if they applied any shipping or discounts.